+++ pyobjc-core/Modules/objc/method-signature.h
@@ -51,7 +51,8 @@ struct _PyObjCMethodSignature {
     unsigned char            null_terminated_array : 1;
     unsigned char            free_result : 1;
     unsigned char            shortcut_signature : 1;
-    unsigned int             shortcut_argbuf_size : 12;
+    unsigned int             shortcut_argbuf_size : 10;
+    unsigned int             shortcut_result_size : 8;
     int16_t                  arrayArg;
     int                      deprecated;
     struct _PyObjC_ArgDescr* rettype;