+++ docs/changelog.rst
@@ -90,6 +90,11 @@ in older Python versions is unchanged except for the effects of general cleanup.
 
 * Convert internal calls into Python to the vectorcall protocol (pyobjc-core)
 
+* Add more optimized vectorcall implementation to :class:`objc.function` and
+  :class:`objc.selector` for simpler callables.
+
+  XXX: To describe wat 'simpler' actually is...
+
 New features
 ............